Fixed put method to create entity if none exists, added delete method
WHY: spec states that put "may also be used to create an Entity with a specific pre-determined ID", meaning that if no entity exists with that id, we must create one. spec also specifies addition of "DELETE" method

HOW: for PUT method, if file already exists, we update and return http::status::ok. if file doesn't exist, we create and return http::status::created. for DELETE method, we ensure entity_name and entity_id are present. Then, we define a new deleteObject(CrudPath path) method for the CrudFileManager. This returns true if the file was successfully deleted, and false if the file doesn't exist/had an error in deletion. If the file was deleted, we return http::status::ok. Else, we return http::status::no_content, indicating that no delete was performed, but that the request was processed successfully

TESTING: Added two tests for crud_file_manager (delete an object, delete a non-existent object), added three tests for crud_handler (delete successfully, delete without entity_id, delete a non-existent entity). Coverage > 94%
Change-Id: I9e718927730505db47e87b9fbaecaf6a2ed63388