Added multithreading to the server with a threadpool.
Currently, the size of the threadpool is set by a macro in _main, but this should probably be configurable in the config file.

Made signal_handler depreciated by handling signals in server. This was to make rejoining the threads possible.

Also fixed corrupted log buffer problems Akiva pointed out.

Also uncommented line in multithread integration test to test
multithread capabilities.

Change-Id: I7ec21d6e22c157ccf42a9ef3606c59a807ab6d35